Bugfix: Show correct home path of the WordPress installation on the plugins admin page.
If WordPress is not installed in the root folder of the domain, a wrong path of the WordPress installation is shown on the admin page of the plugin. This may confuse users about the correct script path.

This problem does not occur, if get_home_path() is used to determine the path of the WordPress installation (see http://codex.wordpress.org/Function_Reference/get_home_path).